requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
firecrawl-py==4.14.0
supabase==2.12.0
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.10.5
email-validator==2.2.0
google-generativeai==0.8.3
resend==2.4.0
pillow==11.0.0
